[refactor] 배송비계산 관련 타입 정리
