+++ libs/hooks/src/lib/queries/useOrderShippingCheck.tsx
@@ -1,4 +1,7 @@
-import { OrderShippingCheckDto } from '@project-lc/shared-types';
+import {
+  OrderShippingCheckDto,
+  ShippingCostByShippingGroupId,
+} from '@project-lc/shared-types';
 import { AxiosError } from 'axios';
 import { useQuery, UseQueryResult } from 'react-query';
 import axios from '../../axios';
@@ -8,15 +11,7 @@ export interface OrderShippingCheck {
 }
 
 /** 배송비정보 타입 */
-export type OrderShippingData = Record<
-  number, // 배송비그룹 id
-  {
-    isShippingAvailable?: boolean;
-    cost: { std: number; add: number } | null; // 기본배송비, 추가배송비
-    items: number[]; // 해당배송비에 포함된 goodsId[]
-  }
->;
-
+export type OrderShippingData = ShippingCostByShippingGroupId;
 export const getOrderShippingCheck = async (
   dto?: OrderShippingCheckDto,
 ): Promise<OrderShippingData> => {
